Cleaned UpdateProjectTask commands mess
