+++ src/core/Codend.Application/Codend.Application.csproj
@@ -16,7 +16,10 @@
         <PackageReference Include="FluentResults" Version="3.15.2"/>
         <PackageReference Include="MediatR" Version="12.1.1"/>
         <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.10" />
+
+
     </ItemGroup>
+
