+++ src/core/Codend.Application/ProjectTasks/Commands/UpdateProjectTask/UpdateBugfixProjectTaskCommand.cs
@@ -1,5 +1,6 @@
 using Codend.Application.Core.Abstractions.Data;
 using Codend.Application.Core.Abstractions.Messaging.Commands;
+using Codend.Application.ProjectTasks.Commands.UpdateProjectTask.Abstractions;
 using Codend.Domain.Entities;
 using Codend.Domain.Repositories;
