+++ src/core/Codend.Domain/Entities/ProjectTask/Abstraction/IProjectTaskUpdater.cs
@@ -3,7 +3,7 @@ using FluentResults;
 
 namespace Codend.Domain.Entities;
 
-public abstract record UpdateProjectTaskProperties
+public abstract record UpdateAbstractProjectTaskProperties
 (
     IShouldUpdate<string> Name,
     IShouldUpdate<string> Priority,
@@ -17,7 +17,7 @@ public abstract record UpdateProjectTaskProperties
 
 public interface IProjectTaskUpdater<TProjectTask, in TUpdateProps>
     where TProjectTask : AbstractProjectTask
-    where TUpdateProps : UpdateProjectTaskProperties
+    where TUpdateProps : UpdateAbstractProjectTaskProperties
 {
     Result<TProjectTask> Update(TUpdateProps properties);
 }
